added RLE support

+++ src/memory.c
@@ -2,7 +2,7 @@
 
 #include "common.h"
 #include "memory.h"
-void* reallocate(void *previous, size_t oldSize, size_t newSize) {
+void* reallocate(void *previous, int oldSize, int newSize) {
   if(newSize == 0) {
     free(previous);
     return NULL;
